Added functionality to copy and delete buttons

+++ lib/gist_clone/gists.ex
@@ -8,6 +8,7 @@ defmodule GistClone.Gists do
 
   alias GistClone.Gists.Gist
   alias GistClone.Gists.SavedGist
+  alias GistClone.Accounts.User
 
   @doc """
   Returns the list of gists.
@@ -70,10 +71,16 @@ defmodule GistClone.Gists do
       {:error, %Ecto.Changeset{}}
 
   """
-  def update_gist(%Gist{} = gist, attrs) do
-    gist
-    |> Gist.changeset(attrs)
-    |> Repo.update()
+  def update_gist(%User{} = user, gist_id, attrs) do
+    gist = Repo.get!(Gist, gist_id)
+
+    if user.id == gist.user_id do
+      gist
+      |> Gist.changeset(attrs)
+      |> Repo.update()
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
@@ -88,8 +95,14 @@ defmodule GistClone.Gists do
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_gist(%Gist{} = gist) do
-    Repo.delete(gist)
+  def delete_gist(%User{} = user, gist_id) do
+    gist = Repo.get!(Gist, gist_id)
+
+    if user.id == gist.user_id do
+      Repo.delete(gist)
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
